refactor based on gnuradio dev guidance
According to willcode4, sources and sinks should inherit from gr::block
Also, the work() function (or general_work() in this case) is expected
to block until the requested number of items are available.

This implementation sets an output_multiple so that it is only called
when a whole packet can be consumed. For formats that support it,
dropped packet detection is implemented by inserted null samples for
each missing packet.